Move Staking test implementation
